+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
-team.md 
+# team.md 
 .vscode/
 .DS_store
