+++ client/src/components/login/Form.js
@@ -1,0 +1,61 @@
+import Button from '@material-ui/core/Button';
+import TextField from '@material-ui/core/TextField';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Checkbox from '@material-ui/core/Checkbox';
+import Grid from '@material-ui/core/Grid';
+import useStyles from '../components/login/Styles';
+
+export default function Form() {
+  const classes = useStyles();
+
+  <div className={classes.paper}>
+    <Typography component="h1" variant="h4">
+      Login to Bill Tracker
+        </Typography>
+    <form className={classes.form} noValidate>
+      <TextField
+        variant="outlined"
+        margin="normal"
+        required
+        fullWidth
+        id="email"
+        label="Email Address"
+        name="email"
+        autoComplete="email"
+        autoFocus
+      />
+      <TextField
+        variant="outlined"
+        margin="normal"
+        color="4B7631"
+        required
+        fullWidth
+        name="password"
+        label="Password"
+        type="password"
+        id="password"
+        autoComplete="current-password"
+      />
+      <FormControlLabel
+        control={<Checkbox value="remember" color="primary" />}
+        label="Remember me"
+      />
+      <Button
+        type="submit"
+        fullWidth
+        variant="contained"
+        color="primary"
+        className={classes.submit}
+      >
+        Sign In
+          </Button>
+      <Grid container>
+        <Grid item>
+          <Link to={process.env.PUBLIC_URL + "/signup"} variant="body2">
+            {"Don't have an account? Sign Up"}
+          </Link>
+        </Grid>
+      </Grid>
+    </form>
+  </div>
+}
