+++ client/src/components/signup/Form.js
@@ -1,0 +1,71 @@
+<div className={classes.paper}>
+  <Typography component="h1" variant="h5">
+    Sign up
+        </Typography>
+  <form className={classes.form} noValidate>
+    <Grid container spacing={2}>
+      <Grid item xs={12} sm={6}>
+        <TextField
+          autoComplete="fname"
+          name="firstName"
+          variant="outlined"
+          required
+          fullWidth
+          id="firstName"
+          label="First Name"
+          autoFocus
+        />
+      </Grid>
+      <Grid item xs={12} sm={6}>
+        <TextField
+          variant="outlined"
+          required
+          fullWidth
+          id="lastName"
+          label="Last Name"
+          name="lastName"
+          autoComplete="lname"
+        />
+      </Grid>
+      <Grid item xs={12}>
+        <TextField
+          variant="outlined"
+          required
+          fullWidth
+          id="email"
+          label="Email Address"
+          name="email"
+          autoComplete="email"
+        />
+      </Grid>
+      <Grid item xs={12}>
+        <TextField
+          variant="outlined"
+          required
+          fullWidth
+          name="password"
+          label="Password"
+          type="password"
+          id="password"
+          autoComplete="current-password"
+        />
+      </Grid>
+    </Grid>
+    <Button
+      type="submit"
+      fullWidth
+      variant="contained"
+      color="primary"
+      className={classes.submit}
+    >
+      Sign Up
+          </Button>
+    <Grid container justify="center">
+      <Grid item>
+        <Link href="#" variant="body2">
+          Already have an account? Sign in
+              </Link>
+      </Grid>
+    </Grid>
+  </form>
+</div>
